removed old stuff and extra new stuff
